feat(mobile): surface unassigned captures in the Inbox, not only the dashboard
A capture caught without a group (A34) is something waiting for you, so it now
shows in the Inbox as well as on the dashboard — the two screens a person opens
to answer "is there anything for me?". Tapping it opens the captures list, same
as the dashboard card.

Extracted the card into a shared `UnassignedCapturesCard` that reads the count
itself and renders nothing when there is none, so both screens drop it in with
no guard of their own. In the Inbox the empty "nothing yet" state stands down
when a capture is waiting above, so the screen never says it is empty while it
is showing a card.
